simple_requirements.txt
streamlit==1.36.0
ultralytics==8.3.15
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4